// EscalafonFrontend/src/app/interfaces/archivo.ts
export interface Archivo {
    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { number } - iArchFolios
     */
    iArchFolios?: number;

    /**
     * @type { string } - cArchExtension
     */
    cArchExtension?: string;

    /**
     * @type { string } - cArchUuid
     */
    cArchUuid?: string;

}