// EscalafonFrontend/src/app/interfaces/v-experiencia-laboral.ts
export interface VExperienciaLaboral {
    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iExpLabSecId
     */
    iExpLabSecId?: number;

    /**
     * @type { string } - cExpLabEntidad
     */
    cExpLabEntidad?: string;

    /**
     * @type { string } - cExpLabCargo
     */
    cExpLabCargo?: string;

    /**
     * @type { string } - cExpLabNumeroDocumento
     */
    cExpLabNumeroDocumento?: string;

    /**
     * @type { Date } - dtExpLabFechaDocumento
     */
    dtExpLabFechaDocumento?: Date;

    /**
     * @type { Date } - dtExpLabFechaInicio
     */
    dtExpLabFechaInicio?: Date;

    /**
     * @type { Date } - dtExpLabFechaFin
     */
    dtExpLabFechaFin?: Date;

    /**
     * @type { string } - cExpLabAnotaciones
     */
    cExpLabAnotaciones?: string;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { string } - cExpLabSecNombre
     */
    cExpLabSecNombre?: string;

    /**
     * @type { string } - cExpLabFuncionesDesempenadas
     */
    cExpLabFuncionesDesempenadas?: string;

    /**
     * @type { number } - iExpLabId
     */
    iExpLabId?: number;

}